Multithresh code updated
The code is in folders, due to many files in each function. Even if the
final artifact is just one file, many papers and test images are need.
Otsu paper was included in multithresh folder. This paper is easy to
find in internet.
An answer_draft.txt to the first version of the multithresh function was
included. Most answer text is ready, need to finish the code.
Some changes in multithres() function, but it is not ready yet.