angl2str.m use round() instead of floor()
To pass degrees fractional part to minutes and seconds, uses the floor()
function to decide if fractional part is zero or not. In the future will
need to improve Octave's round() function to accept more arguments. An
initial expanded round() function, that I call my_round, was implemented
as a private function.