angl2str.m use %02.f to complete leading zero
Instead of check if number is less than 10 to add the leading zero, uses
the format argument in num2str().
When it do not need two digits before point in numbers, then the
private function round2str() is not call.
Tests checking wrong input moved to beginning.
To convert char to a column vector of the same char, it uses a variable
l instead of call ones() many times.