angl2str.m style update to mapping standards
Long lines divided in two or three. It applys to comments, string
concatenations and calculations.
Lines with only comments uses two # instead of one #.
Use of ismember() instead of some strcmp().
Verification of arguments types made before of others verifications.
Function round() was included as an inner function renamed to round_a().
I tested the use of roundn(), but probably there is one bug in that
function.